Salva alterações locais antes do pull

+++ src/routes/atracoes.js
@@ -6,43 +6,50 @@ const db = require("../db");
 router.get("/", async (req, res) => {
   try {
     const r = await db.query("SELECT * FROM  atracao;");
+    if (!r.rowCount) {
+      throw new Error("Nenhuma atração encontrada!")
+    }
     return res.status(200).json(r.rows)
   } catch (error) {
-    return res.status(400).json({ msg: error.message })
+    return res.status(404).json({ msg: error.message })
   }
 })
 
 router.get("/:id", async (req, res) => {
   try {
-    let id = req.params.id
-    const r = await db.query("SELECT * FROM atracao WHERE id = $1", [id])
+    const id = Number(req.params.id);
+    if (!Number.isInteger(id)) {
+      throw new Error("ID inválido!" )
+    }
+    const r = await db.query("SELECT * FROM atracao WHERE id = $1", [id]);
     if (!r.rowCount) {
-      throw new Error("Atracao nao cadastrada")
+      throw new Error("Atração não cadastrada" );
     } else {
-      return res.status(200).json(r.rows)
+      return res.status(200).json(r.rows[0]);
     }
   } catch (error) {
-    return res.status(400).json({ msg: error.message })
+    return res.status(404).json({ msg: error.message });
   }
 })
 
 router.post("/", async (req, res) => {
   try {
-    let nome = req.body.nome || {};
-    let nacionalidade = req.body.nacionalidade || {};
-    let tipo = req.body.tipo|| {};
+    const { nome, nacionalidade, tipo, generoMusical } = req.body || {}
+
     if (!nome || nome.length < 2) {
-      throw new Error("Nome incompleto")
+      throw new Error("Nome incompleto" );
     } else if (!nacionalidade || nacionalidade.length < 2) {
-      throw new Error("Nacionalidade incompleta")
-    }else if (!tipo|| tipo.length < 2) {
-      throw new Error ("Tipo incompleto")
-    }else {
-      const r = await db.query("INSERT INTO atracao (nome, nacionalidade, tipo) VALUES ($1, $2, $3) RETURNING *", [nome, nacionalidade, tipo]);
+      throw new Error("Nacionalidade incompleta" );
+    } else if (!tipo || tipo.length < 2) {
+      throw new Error("Tipo incompleto" );
+    } else if (!Number.isInteger(generoMusical)) {
+      throw new Error("ID do Gênero musical inválido" );
+    } else {
+      const r = await db.query("INSERT INTO atracao (nome, nacionalidade, tipo, genero_musical_id) VALUES ($1, $2, $3, $4) RETURNING *", [nome, nacionalidade, tipo, generoMusical]);
       if (!r.rowCount) {
-        throw new Error("Atração nao adicionada")
+        throw new Error("Atração não adicionada");
       } else {
-        return res.status(200).json(r.rows);
+        return res.status(201).json({msg: "Uhuuul, atração adicionada!", atracao: r.rows[0]});
       }
     }
   } catch (error) {
@@ -52,37 +59,40 @@ router.post("/", async (req, res) => {
 router.put("/:id", async (req, res) => {
   try {
     let id = req.params.id;
-    let nome = req.body.nome || {};
-    let nacionalidade = req.body.nacionalidade || {};
-    let tipo = req.body.tipo|| {};
+    let { nome, nacionalidade, tipo, generoMusical} = req.body || {}
     if (!nome || nome.length < 2) {
       throw new Error("Nome incompleto")
     } else if (!nacionalidade || nacionalidade.length < 2) {
       throw new Error("Nacionalidade incompleta")
-    }else if (!tipo|| tipo.length < 2) {
-      throw new Error ("Tipo incompleto")
+    } else if (!tipo || tipo.length < 2) {
+      throw new Error("Tipo incompleto")
+    }else if (!Number.isInteger(generoMusical)){
+      throw new Error ("ID do gênero musical inválido")
     } else {
-      const r = await db.query("UPDATE atracao SET nome = $1, nacionalidade = $2, tipo = $3 WHERE id = $4 RETURNING *", [nome, nacionalidade, tipo, id]);
+      const r = await db.query("UPDATE atracao SET nome = $1, nacionalidade = $2, tipo = $3, genero_musical_id = $4 WHERE id = $5 RETURNING *", [nome, nacionalidade, tipo,generoMusical, id]);
       if (!r.rowCount) {
-        throw new Error("Atracao nao editada")
+        throw new Error("Atração não editada")
       } else {
-        return res.status(200).json(r.rows)
+        return res.status(200).json({msg: "Uhuul, atração editada!", atracao: r.rows[0]})
       }
     }
   } catch (error) {
     return res.status(400).json({ msg: error.message })
   }
 })
-router.delete ("/:id", async (req, res)=> {
+router.delete("/:id", async (req, res) => {
   try {
-    const id = req.params.id;
-    const r = await db.query ("DELETE FROM atracao WHERE id = $1", [id])
-    if (!r.rowCount) {
-      throw new Error ("Atracao com esse id ja nao existe!")
+    const id = Number(req.params.id);
+    if (!Number.isInteger(id)){
+      throw new Error ("ID da atração inválido")
     }
-    return res.status(200).json ({msg: "Apagado!"})
-  }catch (error) {
-    return res.status(400).json({msg: error.message})
+    const r = await db.query("DELETE FROM atracao WHERE id = $1", [id])
+    if (!r.rowCount) {
+      throw new Error("Atracao com esse id ja nao existe!")
+    }
+    return res.status(200).json({ msg: "Atração apagada!" })
+  } catch (error) {
+    return res.status(400).json({ msg: error.message })
   }
 })
 module.exports = router;
